+++ src/RealEstateInvesting.Application/Auth/DTOs/VerifyWalletRequest.cs
@@ -11,7 +11,7 @@ namespace RealEstateInvesting.Application.Auth.DTOs;
 public class VerifyWalletRequest
 {
     public string WalletAddress { get; set; } = null!;
-    public int ChainId { get; set; }
+    public long ChainId { get; set; }
     public string Signature { get; set; } = null!;
 
     // 🔥 EXACT string that wallet signed
